:bug: fix link & unlink make targets

+++ Makefile
@@ -10,8 +10,8 @@ test:
 
 link:
 	mkdir -p $(GOPATH)/src/github.com/hico-horiuchi
-	ln -s $(CURDIR) $(GOPATH)/src/github.com/hico-horiuchi/ohgibone
+	if [ ! -d $(GOPATH)/src/github.com/hico-horiuchi/ohgibone ]; then ln -s $(CURDIR) $(GOPATH)/src/github.com/hico-horiuchi/ohgibone; fi
 
 unlink:
 	rm $(GOPATH)/src/github.com/hico-horiuchi/ohgibone
-	rmdir $(GOPATH)/src/github.com/hico-horiuchi
+	if [ -z "`ls $(GOPATH)/src/github.com/hico-horiuchi`" ]; then rmdir $(GOPATH)/src/github.com/hico-horiuchi; fi
